Import modifications in response to files being moved.
