+++ src/main/java/use_cases/launch_new_game_use_case/LaunchNewGameInteractor.java
@@ -1,6 +1,7 @@
 package use_cases.launch_new_game_use_case;
 
 import core_entities.game_parts.*;
+import core_entities.game_parts.board.BoardFactory;
 import core_entities.player.ComputerPlayer;
 import core_entities.player.HumanPlayer;
 import core_entities.player.Player;
