+++ src/main/java/use_cases/place_word_refill_user_story/PlaceWordRefillInteractor.java
@@ -1,5 +1,6 @@
 package use_cases.place_word_refill_user_story;
 
+import core_entities.game_parts.board.Board;
 import core_entities.player.Player;
 import core_entities.game_parts.*;
